+++ source/renderer/modules/markdown-editor/hooks/autocomplete.js
@@ -17,7 +17,56 @@ var currentDatabase = null
 var availableDatabases = {
   'tags': [],
   'citekeys': [],
-  'files': []
+  'files': [],
+  'syntaxHighlighting': [
+    { text: 'javascript', displayText: 'JavaScript/Node.JS' },
+    { text: 'json', displayText: 'JSON' },
+    { text: 'typescript', displayText: 'TypeScript' },
+    { text: 'c', displayText: 'C' },
+    { text: 'cpp', displayText: 'C++' },
+    { text: 'csharp', displayText: 'C#' },
+    { text: 'clojure', displayText: 'Clojure' },
+    { text: 'elm', displayText: 'Elm' },
+    { text: 'fsharp', displayText: 'F#' },
+    { text: 'java', displayText: 'Java' },
+    { text: 'kotlin', displayText: 'Kotlin' },
+    { text: 'haskell', displayText: 'Haskell' },
+    { text: 'objectivec', displayText: 'Objective-C' },
+    { text: 'scala', displayText: 'Scala' },
+    { text: 'css', displayText: 'CSS' },
+    { text: 'scss', displayText: 'SCSS' },
+    { text: 'less', displayText: 'LESS' },
+    { text: 'html', displayText: 'HTML' },
+    { text: 'markdown', displayText: 'Markdown' },
+    { text: 'xml', displayText: 'XML' },
+    { text: 'tex', displayText: 'TeX' },
+    { text: 'php', displayText: 'PHP' },
+    { text: 'python', displayText: 'Python' },
+    { text: 'r', displayText: 'R' },
+    { text: 'ruby', displayText: 'Ruby' },
+    { text: 'sql', displayText: 'SQL' },
+    { text: 'swift', displayText: 'Swift' },
+    { text: 'bash', displayText: 'Bash' },
+    { text: 'visualbasic', displayText: 'Visual Basic' },
+    { text: 'yaml', displayText: 'YAML' },
+    { text: 'go', displayText: 'Go' },
+    { text: 'rust', displayText: 'Rust' },
+    { text: 'julia', displayText: 'Julia' },
+    { text: 'turtle', displayText: 'Turtle' },
+    { text: 'sparql', displayText: 'SparQL' },
+    { text: 'verilog', displayText: 'Verilog' },
+    { text: 'systemverilog', displayText: 'SystemVerilog' },
+    { text: 'vhdl', displayText: 'VHDL' },
+    { text: 'tcl', displayText: 'TCL' },
+    { text: 'scheme', displayText: 'Scheme' },
+    { text: 'clisp', displayText: 'Common Lisp' },
+    { text: 'powershell', displayText: 'Powershell' },
+    { text: 'smalltalk', displayText: 'Smalltalk' },
+    { text: 'dart', displayText: 'Dart' },
+    { text: 'toml', displayText: 'TOML/INI' },
+    { text: 'docker', displayText: 'Dockerfile' },
+    { text: 'diff', displayText: 'Diff' }
+  ]
 }
 
 const CodeMirror = require('codemirror')
@@ -26,9 +75,16 @@ module.exports = {
   'autocompleteHook': (cm) => {
     // Listen to change events
     cm.on('change', (cm, changeObj) => {
+      if (autocompleteStart !== null && currentDatabase !== null) {
+        // We are currently autocompleting something, let's finish that first.
+        return
+      }
+
       let autocompleteDatabase = shouldBeginAutocomplete(cm, changeObj)
 
-      if (autocompleteDatabase === undefined) return
+      if (autocompleteDatabase === undefined) {
+        return
+      }
 
       autocompleteStart = Object.assign({}, cm.getCursor())
       currentDatabase = availableDatabases[autocompleteDatabase]
@@ -50,16 +106,28 @@ module.exports = {
 }
 
 function shouldBeginAutocomplete (cm, changeObj) {
-  // The easiest are citekeys
-  if (changeObj.text[0] === '@') return 'citekeys'
+  const cursor = cm.getCursor()
+  const line = cm.getLine(cursor.line)
 
-  // Now we need some more stuff
-  let cursor = cm.getCursor()
-  let line = cm.getLine(cursor.line)
+  // Can we begin citekey autocompletion?
+  // A valid citekey position is: Beginning of the line (citekey without square
+  // brackets), after a square bracket open (regular citation without prefix),
+  // or after a space (either a standalone citation or within square brackets
+  // but with a prefix).
+  if (
+    changeObj.text[0] === '@' &&
+    (cursor.ch === 1 || [ ' ', '[' ].includes(line.charAt(cursor.ch - 2)))
+  ) {
+    return 'citekeys'
+  }
 
   // Can we begin tag autocompletion?
-  if (changeObj.text[0] === '#' &&
-    (cursor.ch === 1 || line.charAt(cursor.ch - 2) === ' ')) return 'tags'
+  if (
+    changeObj.text[0] === '#' &&
+    (cursor.ch === 1 || line.charAt(cursor.ch - 2) === ' ')
+  ) {
+    return 'tags'
+  }
 
   // Can we begin file autocompletion?
   let linkStart = cm.getOption('zettlr').zettelkasten.linkStart
@@ -73,6 +141,24 @@ function shouldBeginAutocomplete (cm, changeObj) {
 
   if (linkStartRange === linkStart) return 'files'
 
+  // Now check for syntax highlighting
+  if (
+    (line.startsWith('```') || line.startsWith('~~~')) &&
+    cursor.ch === 3 // We should only begin autocomplete immediately after the delimiters
+  ) {
+    // First line means it's definitely the beginning of the block
+    if (cursor.line === 0) {
+      return 'syntaxHighlighting'
+    }
+    // Check if the mode on the line *before* is still Markdown
+    let modeLineBefore = cm.getModeAt({ line: cursor.line - 1, ch: 0 })
+    if (modeLineBefore.name === 'markdown') {
+      // If our own line starts with a codeblock, but the line before is
+      // still in Markdown mode, this means we have the beginning of a codeblock.
+      return 'syntaxHighlighting'
+    }
+  }
+
   return undefined // Nothing to do for us here
 }
 
@@ -82,7 +168,7 @@ function hint (cm, opt) {
     'list': Object.keys(currentDatabase).filter((key) => {
       // First search the ID. Second, search the displayText, if available.
       // Third: return false if nothing else has matched.
-      if (currentDatabase[key].text.toLowerCase().indexOf(term) === 0) return true
+      if (currentDatabase[key].text.toLowerCase().indexOf(term) >= 0) return true
       if (currentDatabase[key].hasOwnProperty('displayText') && currentDatabase[key].displayText.toLowerCase().indexOf(term) >= 0) return true
       return false
     })
@@ -138,6 +224,21 @@ function hint (cm, opt) {
         cm.replaceSelection(prefix + text)
       } else if (linkEndMissing) {
         cm.replaceSelection(end) // Add the link ending
+      }
+    } else if (currentDatabase === availableDatabases['syntaxHighlighting']) {
+      // In the case of an accepted syntax highlighting, we can assume the user
+      // has manually begun writing a code block, so we are probably right
+      // to assume that the user would think it's nice if we also add the
+      // closing part of the code block and set the cursor in the middle of the
+      // newly rendered codeblock.
+      const line = cm.getLine(autocompleteStart.line)
+      if (line.startsWith('```')) {
+        // completion.text += '\n\n```'
+        cm.replaceSelection('\n\n```')
+        cm.setCursor({ line: autocompleteStart.line + 1, ch: 0 })
+      } else if (line.startsWith('~~~')) {
+        cm.replaceSelection('\n\n```')
+        cm.setCursor({ line: autocompleteStart.line + 1, ch: 0 })
       }
     }
     autocompleteStart = null